discrete_control_requirements.txt
absl-py==1.0.0
ale-py==0.7.5
astunparse==1.6.3
atari-py==0.2.9
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backcall==0.2.0
cached-property==1.5.2
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
chex==0.1.3
clang==5.0
click==8.1.3
cloudpickle==2.0.0
cycler==0.11.0
decorator==5.1.1
dm-haiku==0.0.6
dm-tree==0.1.7
docker-pycreds==0.4.0
dopamine-rl==4.0.2
flatbuffers==2.0
flax==0.4.2
future==0.18.2
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.1
gym==0.23.1
gym-notices==0.0.6
h5py==3.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipdb==0.13.9
ipython==7.33.0
jax==0.3.13
jaxlib==0.3.10+cuda11.cudnn805
jedi==0.18.1
jmp==0.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.4.2
matplotlib-inline==0.1.3
msgpack==1.0.3
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
optax==0.1.2
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyglet==1.4.11
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-probability==0.16.0
termcolor==1.1.0
tf-slim==1.1.0
toml==0.10.2
toolz==0.11.2
tqdm==4.64.0
traitlets==5.2.1.post0
typing-extensions==4.2.0
urllib3==1.26.9
wandb==0.12.16
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
